+++ packages/react/webpack.config.js
@@ -2,23 +2,25 @@ const buildConfig = require('../../scripts/webpackFactory');
 
 const lib = {
   entry: {
-    'yubaba-react': './src/index',
+    'react-yubaba': './src/index',
   },
   externals: {
     react: {
       commonjs: 'react',
+      commonjs2: 'react',
       amd: 'react',
       root: 'React',
     },
     'yubaba-core': {
       commonjs: 'yubaba-core',
+      commonjs2: 'yubaba-core',
       amd: 'yubaba',
       root: 'yubaba',
     },
   },
   path: './dist',
   filename: '[name].js',
-  library: 'yubabaReact',
+  library: 'reactYubaba',
 };
 
 const makeTestPage = (name) => ({